+++ dse_do_utils/scenariomanager.py
@@ -8,6 +8,8 @@
 # -----------------------------------------------------------------------------------
 import os
 import glob
+import pathlib
+import zipfile
 
 import docplex
 import pandas as pd
@@ -514,19 +516,42 @@ class ScenarioManager(object):
     # -----------------------------------------------------------------
     # Read and write from/to Excel - value added-functions
     # -----------------------------------------------------------------
-
     def load_data_from_excel(self, excel_file_name: str) -> InputsOutputs:
         """Load data from an Excel file located in the `datasets` folder of the root directory.
         Convenience method.
         If run not on WS, requires the `root_dir` property passed in the ScenarioManager constructor
         """
-        # root_dir = self.get_root_directory()
-        datasets_dir = self.get_data_directory()
-        excel_file_path = os.path.join(datasets_dir, excel_file_name + '.xlsx')
-        xl = pd.ExcelFile(excel_file_path)
+        if pathlib.Path(excel_file_name).suffix == '.xlsx':
+            file_name = excel_file_name
+        else:
+            file_name = excel_file_name + '.xlsx'
+
+        if self.platform == Platform.CPDaaS:
+            # For CPDaaS only: file doesn't exist in /home/wsuser/work/. We have to get it.
+            file = self.project.get_file(file_name)
+            file.seek(0)
+            xl = pd.ExcelFile(file)
+        else:
+            datasets_dir = self.get_data_directory()
+            excel_file_path = os.path.join(datasets_dir, file_name)
+            xl = pd.ExcelFile(excel_file_path)
+
         # Read data from Excel
         self.inputs, self.outputs = ScenarioManager.load_data_from_excel_s(xl)
         return self.inputs, self.outputs
+
+    # def load_data_from_excel(self, excel_file_name: str) -> InputsOutputs:
+    #     """Load data from an Excel file located in the `datasets` folder of the root directory.
+    #     Convenience method.
+    #     If run not on WS, requires the `root_dir` property passed in the ScenarioManager constructor
+    #     """
+    #     # root_dir = self.get_root_directory()
+    #     datasets_dir = self.get_data_directory()
+    #     excel_file_path = os.path.join(datasets_dir, excel_file_name + '.xlsx')
+    #     xl = pd.ExcelFile(excel_file_path)
+    #     # Read data from Excel
+    #     self.inputs, self.outputs = ScenarioManager.load_data_from_excel_s(xl)
+    #     return self.inputs, self.outputs
 
     def write_data_to_excel(self, excel_file_name: str = None, copy_to_csv: bool = False) -> None:
         """Write inputs and/or outputs to an Excel file in datasets.
@@ -550,14 +575,17 @@ class ScenarioManager(object):
                 raise ValueError("The argument excel_file_name can only be 'None' if both the model_name '{}' and the scenario_name '{}' have been specified.".format(self.model_name, self.scenario_name))
 
         # root_dir = self.get_root_directory()
-        # Save the regular Excel file:
+        # Save the Excel file:
+        if pathlib.Path(excel_file_name).suffix != '.xlsx':
+            excel_file_name = excel_file_name + '.xlsx'
+
         data_dir = self.get_data_directory()
-        excel_file_path_1 = os.path.join(data_dir, excel_file_name + '.xlsx')
+        excel_file_path_1 = os.path.join(data_dir, excel_file_name)
         writer_1 = pd.ExcelWriter(excel_file_path_1, engine='xlsxwriter')
         ScenarioManager.write_data_to_excel_s(writer_1, inputs=self.inputs, outputs=self.outputs)
         writer_1.save()
 
-        self.add_file_as_data_asset(excel_file_path_1, excel_file_name + '.xlsx')
+        self.add_file_as_data_asset(excel_file_path_1, excel_file_name)
 
         # if self.platform == Platform.CPDaaS:
         #     self.add_data_file_using_project_lib(excel_file_path_1, excel_file_name + '.xlsx')
@@ -575,18 +603,38 @@ class ScenarioManager(object):
         #     os.rename(excel_file_path_2, csv_excel_file_path_2)
         return excel_file_path_1
 
-    def add_file_as_data_asset(self, file_path: str, asset_name:str = None):
+    def add_file_as_data_asset(self, file_path: str, asset_name: str = None):
         """Register an existing file as a data asset in CPD.
 
         :param file_path: full path of the file
         :param asset_name: name of asset. If None, will get the name from the file
         :return:
         """
-        ScenarioManager.add_file_as_data_asset_s(file_path, asset_name, self.platform)
+        if asset_name is None:
+            asset_name = os.path.basename(file_path)
+
+        if self.platform in [Platform.CPD40]:
+            self.add_data_file_using_ws_lib_s(file_path, asset_name)
+        elif self.platform in [Platform.CPD25, Platform.CPDaaS]:
+            self.add_data_file_using_project_lib(file_path, asset_name)
+        else:  # i.e Local: do not register as data asset
+            pass
+
+    # def add_file_as_data_asset(self, file_path: str, asset_name:str = None):
+    #     """Register an existing file as a data asset in CPD.
+    #
+    #     :param file_path: full path of the file
+    #     :param asset_name: name of asset. If None, will get the name from the file
+    #     :return:
+    #     """
+    #     ScenarioManager.add_file_as_data_asset_s(file_path, asset_name, self.platform)
+
+
 
     @staticmethod
     def add_file_as_data_asset_s(file_path: str, asset_name: str = None, platform: Platform = None):
         """Register an existing file as a data asset in CPD.
+        VT 2022-01-21: this method is incorrect for CPDaaS. Should use project_lib.
 
         :param file_path: full path of the file
         :param asset_name: name of asset. If None, will get the name from the file
@@ -984,3 +1032,21 @@ class ScenarioManager(object):
         # elif self.platform == Platform.CPD25:
         #     self.add_data_file_using_project_lib(lp_file_path, lp_file_name)
         return lp_file_path
+
+    def insert_scenarios_from_zip(self, filepath: str):
+        """Insert (or replace) a set of scenarios from a .zip file into the DO Experiment.
+        Zip is assumed to contain one or more .xlsx files. Others will be skipped.
+        Name of .xlsx file will be used as the scenario name."""
+        with zipfile.ZipFile(filepath, 'r') as zip_file:
+            for info in zip_file.infolist():
+                scenario_name = pathlib.Path(info.filename).stem
+                file_extension = pathlib.Path(info.filename).suffix
+                if file_extension == '.xlsx':
+                    #                 print(f"file in zip : {info.filename}")
+                    xl = pd.ExcelFile(zip_file.read(info))
+                    self.inputs, self.outputs = ScenarioManager.load_data_from_excel_s(xl)
+                    self.print_table_names()
+                    self.write_data_into_scenario_s(self.model_name, scenario_name, self.inputs, self.outputs, self.template_scenario_name)
+                    print(f"Uploaded scenario: '{scenario_name}' from '{info.filename}'")
+                else:
+                    print(f"File '{info.filename}' in zip is not a .xlsx. Skipped.")